# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Istaroth Web API",
    "description": "API for the Istaroth Web",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/library/hierarchy": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Get Library Hierarchy",
        "description": "Get the full library hierarchy: every category's document tree.\n\nCategories with a dedicated builder (quests, hangouts) return their pre-baked\nmulti-level tree; any other category returns a flat, depth-1 list of file\nleaves synthesized from the manifest. The corpus is immutable per deployment,\nso the response carries an ETag for free conditional reloads.",
        "operationId": "get_library_hierarchy_api_library_hierarchy_get",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language code (CHS, ENG)",
              "title": "Language"
            },
            "description": "Language code (CHS, ENG)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryHierarchyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/file/{category}/{id}": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Get File",
        "description": "Get full text content of a file by category and id.",
        "operationId": "get_file_api_library_file__category___id__get",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language code (CHS, ENG)",
              "title": "Language"
            },
            "description": "Language code (CHS, ENG)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/proper-nouns": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Get Proper Nouns",
        "description": "Get the static curated proper-noun list for a language.\n\nEmpty when no list ships for the language (e.g. ENG). The frontend uses this\nas a fast fallback while per-file on-the-fly extraction is in flight.",
        "operationId": "get_proper_nouns_api_library_proper_nouns_get",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language code (CHS, ENG)",
              "title": "Language"
            },
            "description": "Language code (CHS, ENG)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProperNounsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/file/{category}/{id}/proper-nouns": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Get File Proper Nouns",
        "description": "Extract a single file's proper nouns for highlighting.\n\nRuns an LLM over the exact rendered content on the fly (cached by content\nhash). Only CHS is supported; ENG returns an empty list.",
        "operationId": "get_file_proper_nouns_api_library_file__category___id__proper_nouns_get",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language code (CHS, ENG)",
              "title": "Language"
            },
            "description": "Language code (CHS, ENG)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProperNounsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/retrieve": {
      "post": {
        "tags": [
          "library"
        ],
        "summary": "Retrieve Library",
        "description": "Retrieve library documents using BM25 or hybrid semantic search.",
        "operationId": "retrieve_library_api_library_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryRetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryRetrieveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HierarchyNode": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "children": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/HierarchyNode"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Children"
          },
          "file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          },
          "toc_eligible": {
            "type": "boolean",
            "title": "Toc Eligible"
          }
        },
        "type": "object",
        "required": [
          "key",
          "title",
          "children",
          "file_id",
          "toc_eligible"
        ],
        "title": "HierarchyNode",
        "description": "One node in a browsable document hierarchy.\n\nA node is either a group (``children`` set) or a leaf (``file_id`` set, a\nviewable file). ``title`` is the resolved display label."
      },
      "LibraryCategoryHierarchy": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/HierarchyNode"
            },
            "type": "array",
            "title": "Nodes"
          }
        },
        "type": "object",
        "required": [
          "category",
          "nodes"
        ],
        "title": "LibraryCategoryHierarchy",
        "description": "One category's document tree within the full library hierarchy."
      },
      "LibraryFileInfo": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "relative_path": {
            "type": "string",
            "title": "Relative Path"
          }
        },
        "type": "object",
        "required": [
          "category",
          "title",
          "id",
          "relative_path"
        ],
        "title": "LibraryFileInfo",
        "description": "File information with parsed components."
      },
      "LibraryFileResponse": {
        "properties": {
          "file_info": {
            "$ref": "#/components/schemas/LibraryFileInfo"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "file_info",
          "content"
        ],
        "title": "LibraryFileResponse",
        "description": "Response model for library file content endpoint."
      },
      "LibraryHierarchyResponse": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/LibraryCategoryHierarchy"
            },
            "type": "array",
            "title": "Categories"
          }
        },
        "type": "object",
        "required": [
          "categories"
        ],
        "title": "LibraryHierarchyResponse",
        "description": "Full library hierarchy: every category's document tree, in display order."
      },
      "LibraryRetrieveRequest": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "k": {
            "type": "integer",
            "title": "K"
          },
          "semantic": {
            "type": "boolean",
            "title": "Semantic",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "language",
          "query",
          "k"
        ],
        "title": "LibraryRetrieveRequest",
        "description": "Request model for library retrieval endpoint."
      },
      "LibraryRetrieveResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/LibraryRetrieveResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "query",
          "results"
        ],
        "title": "LibraryRetrieveResponse",
        "description": "Response model for library retrieval endpoint."
      },
      "LibraryRetrieveResult": {
        "properties": {
          "file_info": {
            "$ref": "#/components/schemas/LibraryFileInfo"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "file_info",
          "snippet",
          "score"
        ],
        "title": "LibraryRetrieveResult",
        "description": "Single retrieved library document with snippet."
      },
...
      "ProperNounsResponse": {
        "properties": {
          "nouns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nouns"
          }
        },
        "type": "object",
        "required": [
          "nouns"
        ],
        "title": "ProperNounsResponse",
        "description": "Response model for the proper-nouns endpoints.\n\n``nouns`` are proper nouns the frontend highlights: either the static curated\nlist for a language or those extracted on the fly from a single file's\ncontent (empty for languages without support, e.g. ENG)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}